pyproject.toml
[tool.poetry]
name = "mashup"
version = "0.1.0"
description = ""
authors = ["이한영 <dev@lhy.kr>"]

[tool.poetry.dependencies]
python = "^3.7"

# Django
django = "<3.0"
django-cors-headers = "^3.1.1"
django-dbbackup = "<3.2"
django-extensions = "^2.2.5"
django-filter = "^2.2.0"
django-modeladmin-reorder = "^0.3.1"
django-oauth-toolkit = "^1.2.0"
django-phonenumber-field = "^3.0.1"
django-push-notifications = "^2.0.0"
django-safedelete = "^0.5.4"
django-secrets-manager = "^0.1.13"
django-simple-history = "^2.8.0"
django-storages = "^1.7.2"

# DRF
djangorestframework = "^3.10.3"
djangorestframework-api-key = "^1.4.1"
djangorestframework-camel-case = "^1.1.2"
django-rest-auth = "^0.9.5"
drf-writable-nested = "^0.5.1"
drf-yasg = "^1.17.0"

# Others
boto3 = "^1.10.14"
jinja2 = "^2.10.3"
pillow = "^6.2.1"
phonenumberslite = "==8.10.21"
psycopg2-binary = "^2.8.4"

# Test
pytest-cov = "^2.8.1"
pytest-django = "^3.7.0"
model_bakery = "^1.0.1"

# Deploy
sentry-sdk = "^0.13.2"
gunicorn = "^20.0.0"
python-dotenv = "^0.12.0"
inflection = "^0.3.1"

[tool.poetry.dev-dependencies]
django-debug-toolbar = "^2.0"
notebook = "^6.0.2"
openpyxl = "^3.0.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
